// src/urn.d.ts
export declare enum BindingNamespace {
    Redirect = "urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Redirect",
    Post = "urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST",
    SimpleSign = "urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST-SimpleSign",
    Artifact = "urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Artifact"
}
export declare enum MessageSignatureOrder {
    STE = "sign-then-encrypt",
    ETS = "encrypt-then-sign"
}
export declare enum StatusCode {
    Success = "urn:oasis:names:tc:SAML:2.0:status:Success",
    Requester = "urn:oasis:names:tc:SAML:2.0:status:Requester",
    Responder = "urn:oasis:names:tc:SAML:2.0:status:Responder",
    VersionMismatch = "urn:oasis:names:tc:SAML:2.0:status:VersionMismatch",
    AuthFailed = "urn:oasis:names:tc:SAML:2.0:status:AuthnFailed",
    InvalidAttrNameOrValue = "urn:oasis:names:tc:SAML:2.0:status:InvalidAttrNameOrValue",
    InvalidNameIDPolicy = "urn:oasis:names:tc:SAML:2.0:status:InvalidNameIDPolicy",
    NoAuthnContext = "urn:oasis:names:tc:SAML:2.0:status:NoAuthnContext",
    NoAvailableIDP = "urn:oasis:names:tc:SAML:2.0:status:NoAvailableIDP",
    NoPassive = "urn:oasis:names:tc:SAML:2.0:status:NoPassive",
    NoSupportedIDP = "urn:oasis:names:tc:SAML:2.0:status:NoSupportedIDP",
    PartialLogout = "urn:oasis:names:tc:SAML:2.0:status:PartialLogout",
    ProxyCountExceeded = "urn:oasis:names:tc:SAML:2.0:status:ProxyCountExceeded",
    RequestDenied = "urn:oasis:names:tc:SAML:2.0:status:RequestDenied",
    RequestUnsupported = "urn:oasis:names:tc:SAML:2.0:status:RequestUnsupported",
    RequestVersionDeprecated = "urn:oasis:names:tc:SAML:2.0:status:RequestVersionDeprecated",
    RequestVersionTooHigh = "urn:oasis:names:tc:SAML:2.0:status:RequestVersionTooHigh",
    RequestVersionTooLow = "urn:oasis:names:tc:SAML:2.0:status:RequestVersionTooLow",
    ResourceNotRecognized = "urn:oasis:names:tc:SAML:2.0:status:ResourceNotRecognized",
    TooManyResponses = "urn:oasis:names:tc:SAML:2.0:status:TooManyResponses",
    UnknownAttrProfile = "urn:oasis:names:tc:SAML:2.0:status:UnknownAttrProfile",
    UnknownPrincipal = "urn:oasis:names:tc:SAML:2.0:status:UnknownPrincipal",
    UnsupportedBinding = "urn:oasis:names:tc:SAML:2.0:status:UnsupportedBinding"
}
// ...
declare const wording: {
    urlParams: {
        samlRequest: string;
        samlResponse: string;
        logoutRequest: string;
        logoutResponse: string;
        sigAlg: string;
        signature: string;
        relayState: string;
    };
    binding: {
        redirect: string;
        post: string;
        simpleSign: string;
        artifact: string;
        soap: string;
    };
    certUse: {
        signing: string;
        encrypt: string;
    };
    metadata: {
        sp: string;
        idp: string;
    };
};